Extract lifes counter and questions counter to common area
