+++ front/src/controls/lifes.css
