+++ front/src/controls/remaining_questions.css
