+++ front/src/controls/remaining_questions.tsx
@@ -7,6 +7,8 @@ export interface Properties {
 }
 
 export default function RemainingQuestionsComponent(props: Properties) {
+  const [containerClass, setContainerClass] = useState<string>('questions-display');
+
   const questionsRemaining = useMemo(() => {
     return props.totalQuestions - props.answeredQuestions;
   }, [props.totalQuestions, props.answeredQuestions]);
@@ -19,9 +21,6 @@ export default function RemainingQuestionsComponent(props: Properties) {
     setContainerClass('questions-display question-switched');
     setTimeout(() => setContainerClass('questions-display'), 1000);
   }, [questionsRemaining]);
-
-  const [containerClass, setContainerClass] =
-    useState<string>('questions-display');
 
   return <div className={containerClass}>{questionsRemaining}</div>;
 }
